requirements.txt
pytest~=8.3.4
selene==2.0.0rc7
allure-pytest
allure-python-commons~=2.13.5
selenium~=4.27.1
python-dotenv~=1.0.1